Fix for incorrect import in ui
