+++ ui/ui-app/src/app/components/modals/LabelsFormGroup.tsx
@@ -9,6 +9,7 @@ export type ArtifactLabel = {
     valueValidated: "success" | "warning" | "error" | "default";
 }
 
+
 /**
  * Labels
  */
