--- ui/ui-app/src/app/components/modals/labelsToList.function.ts
@@ -1,12 +1,0 @@
-import { ArtifactLabel } from "@app/pages";
-
-export function labelsToList(labels: { [key: string]: string|undefined }): ArtifactLabel[] {
-    return Object.keys(labels).filter((key) => key !== undefined).map(key => {
-        return {
-            name: key,
-            value: labels[key],
-            nameValidated: "default",
-            valueValidated: "default"
-        };
-    });
-}
